lib/make.sh: Support $tool prefix for runtestflags in make_check()
When running only a subset of the tests, GCC needs special care,
because it makes use of runtest's --tool option, when looking for
Expect library functions.  Unlike other toolchain components such as
GDB, it does not support RUNTESTFLAGS=path/to/file.exp with plain
'make check': it needs something like 'make check-g++' or
'make-target-libstdc++' for that.

This patch adds support for --set runtestflags=tool:path/to/file.exp,
and uses $tool to select the right make check command and top-level
directory.

This is a bit tedious because we'd need to parse $runtestflags[@]
several times and maintain several lists (dirs, check_targets,
runtestflags), but since this change is only required by GCC, we add
the new handling for GCC only, leaving the simple existing behaviour
to other components.

Change-Id: I889da8661cc2a93e8c12cbed22221f4e371ad000