lib/make.sh (make_check): Mention failures that cause new testsuite runs
If more than two testsuite runs are needed, it's because a test failed in
the second run but it either passed or wasn't executed in the first
testsuite run (and it's also not in the expected failures list).

It's interesting to observe what those failures are, so print them out.

Suggested-by: Maxim Kuvyrkov <maxim.kuvyrkov@linaro.org>

Change-Id: I05d3c49f02622d1750e3a0ce480cdb72a1dca890