config/gdb.conf: Disable Python on MinGW builds
When ABE's config for GDB was updated to GDB 13, the CI job
tcwg-x86_64-build,canadian_cross_build_gcc5 failed to cross-build GDB for
i686-w64-mingw32 because it was trying to use Python 2.x but GDB 13 dropped
support for it.

Attempting to provide an updated Python 3.x built with MinGW is difficult
because Python doesn't support building with the MinGW toolchain¹. There's
a fork from the MSYS2 community adding this support², but using it to
cross-build Python 3.10.10 hits a failure during the installation step with
some PyPI internal error.

Since the point of the CI job is to test ABE's ability to do a Canadian
cross build and we don't actually need to provide GDB for Windows, just
take the easier route and disable Python support when cross-building with
MinGW.

¹ https://github.com/python/cpython/issues/89701
² https://github.com/msys2-contrib/cpython-mingw

Change-Id: Iddd347966ab2b7e416c66a5ccf21c552333ecf09