config/gdb.conf: Point to GMP and MPFR in the prefix directory for MinGW builds
Up to version 13, GDB had its own configure check for GMP and MPFR which
was able to find them in the prefix directory. We relied on this when
building with MinGW.

In the master branch, GDB was changed to use the GMP and MPFR checks in
root's configure.ac but they don't find the libraries in the prefix
directory anymore.

To fix the resulting break of our MinGW build, use specific flags to point
the configure script to GMP and MPFR.

Change-Id: I59dbdb0baa625f2316227b50d0475858a80c2196