gcc: Disable go language
The gotools testsuite does not use dejagnu, and the .sum tests results
do not follow the exact same format as other testsuites, which
confuses validate_failures.py when it tries to parse gotools.sum.

For the moment, it's sufficient to disable go for the sake of
improving our CI stability.

Change-Id: I5882d8c9496926980e75869dff7927afbefe4f20