config/gdb.conf: Add --with-additional-debug-dirs option
Add configure option "--with-additional-debug-dirs=/usr/lib/debug" so that
GDB can find the distro's debug info in /usr/lib/debug. This fixes Ada
tests for tcwg_gnu_native_check_gdb on arm, which use the distro's compiler
and thus need to find the distro's ld.so debug symbols.

At the moment this option is implemented by a patch applied locally, but
it has been submitted upstream.

Change-Id: I08470dd5253083642b64d4355d1d49b94af8d49c