Add --prefix option
Allow defining the prefix directory where the built component will be
installed. This is passed directly to configure.

The purpose is to allow building a GDB with prefix set to /usr, so that it
can find the distro's debug info for ld.so. This is needed for arm builds
so that GDB can correctly recognise Thumb code in ld.so.

Change-Id: I972aacb2cf65a72e57543aa6bbb370cf1055175c