Add --send-results-filter option
This option is meant to be used with GCC's contrib/test_summary to
send validation results to gcc-testresults mailing-list.

Instead of emailing the results directly from ABE, we want to let
Jenkins handle this step, so this option provides a way of "parsing"
the email to transform it into a format suitable for our CI.

The CI scripts can provide a filter whose interface would be:
cat BODY | filter -s SUBJECT RECIPIENT

For instance, it can store BODY, SUBJECT and RECIPIENT in different
files, and make them available to Jenkins.

Change-Id: I204900945433dac046815267e6993e61fa5ef39b