[WIP] lib/make.sh (make_check): Prepare support for various output buffering workarounds
Since GCC's Makefiles have many check-* targets, it is not practicable
to add a single check-read1 like GDB does.  Instead, we could create a
'read1' rule in the top-level Makefile to build read1.so and
expect-read1, like what GDB can do.  Then, the usual check-* targets
can be used, and we can override the READ1 environment variable as
needed (as is already supported by GDB).

This patch partially reverts a previous patch which added READ1
support for GDB, and adds the preliminary call to 'make read1' as
needed.

For GCC, we still default to use_read1=false until the corresponding
GCC patch is merged.

For GCC (and maybe for GDB too, later), we could also consider making
expect-read1 etc... part of ABE, and build it outside of GCC/GDB.

Finally, in addition to READ1, we can think about using coreutils'
stdbuf.

Change-Id: I184a486c7101550ad3cd0de12ddb7fd26a8deb7f