lib/make.sh (make_check): Support various output buffering workarounds
Similar to GDB's "READ1" mode, I experimented with several buffering
workarounds for GCC.

Since GCC's Makefiles have many check-* targets, it is not practicable
to add a single check-read1 like GDB does.  Instead, we could create a
'read1' rule in the top-level Makefile to build read1.so and
expect-read1, like what GDB can do.  Then, the usual check-* targets
can be used, and we can override the READ1 environment variable as
needed (as is already supported by GDB).  Note that this requires a
GCC patch which has not been submitted yet (and may never be, not sure
how useful/usable it would be).

This patch partially reverts a previous patch which added READ1
support for GDB, and adds the preliminary call to 'make read1' as
needed.

In addition to "READ!" more, this patch adds support for coreutils'
stdbuf-based wrappers, provided as helper scripts.  They can be used
by overloading the EXPECT variable when calling 'make check'.  A GCC
patch to fully allow that will be submitted.

The advantage of such wrappers is that the machinery is kept part of
ABE.

Change-Id: I184a486c7101550ad3cd0de12ddb7fd26a8deb7f