lib/make.sh (make_check): Don't use baseline to build flaky tests list
We remove old results from the database, which also removes the flaky
entries they provided. This causes our CI to forget previously detected
flaky tests.

Fix this by making each job record all flaky tests that it detects.  So
even if it is later removed from the results database, other results will
also have the same flaky entries.  This has the additional advantage of
recording how often each flaky entry occurs.

We do this by not using the baseline when generating the lists of flaky
test failures and flaky test passes detected in each testsuite run.

We still use the baseline to decide whether to run another try of the
testsuite, so this change won't impact runtime cost.

With this change we always need to record the $new_flaky tests, except in
$try 0, so a bit of reorganisation of the code at the end of the loop is
needed.

Change-Id: Idf2ec69af9d613511b743f871dc9abfe32e7b977