lib/make.sh (make_check): Copy glibc required gcc libraries
A recent glibc change (1d5024f4f05) made libsupport (which is used
by all tests) to be built with ‘-fexceptions' and '
-fasynchronous-unwind-tables’.  Different than aarch64 and x86,
these options make arm tests require libgcc_so symbols to enable
unwind support ((__aeabi_unwind_cpp_pr0/__aeabi_unwind_cpp_pr1).

This extra required library is handled by glibc test-container,
which is responsible to setup the container tests (all containers
tests now fail due to the missing library).

The easier solution would be to copy both libgcc_s.so.1 and libstdc++.so.6
to the top-level build folder.  It also makes the glibc testing less
dependent on the system configuration (for both libgcc_s.so, which is
used on cancellation and backtrace tests, and libstdc++.so).

Change-Id: I53fb3cfb49fdcdfd629075026219221290abbf60