Add READ1 lib and tool to ABE
Instead of relying on each of the the projects to include support for
READ1, it seems simpler to build it once with ABE, then point to it
when running each project's testsuite.  We could also think about
adding to Dejagnu instead, but adding it to ABE seems easier for the
time being.

This patch mimics what GDB does, and actually copies read1.c from GDB.
We now rely on Makefile.in to expand the appripriate paths at
configure time, so that the expect-read1 wrapper is able to find the
read1.so shared lib in ${abe_top}.

Change-Id: I3500eb714e1a05f5ff99bd73b241812e8879ef26