Add support for --enable maintainer_mode
For components binutils, gcc, gdb, gdbserver, glibc, newlib this flags
adds --enable-maintainer-mode to the configure flags.

Note that this expacts the appropriate versions of autoconf and
automake to be in $PATH (autoconf-2.69 and automake-1.15.1 at the
moment).

Tested by applying a binutils patch series which requires some files
to be regenerated, otherwise there are many failures when running the
testsuite: we do see the failures without this flag, and they
disappear with this flag.

Change-Id: Iadac4c1905761857f976c7b0aa228c4ac1e8c7d5