make.sh: Do not override BUILD_INFO nor MAKEINFO with --disable make_docs
GCC's top-level Makefile does not propagate BUILD_INFO to sub-makes,
instead BUILD_INFO is defined at configure time depending on the
availability of a suitable version of makeinfo (so overriding
MAKEINFO=echo like we used to do is sufficient).

However, it seems cleaner to define a fake makeinfo wrapper outside
ABE (like we do in CI), and don't try to play tricks with GCC's
Makefiles.

With this patch, make_docs only enables/disables building the optional
versions of the documentations (eg HTML).

Change-Id: I56652e5048e2898a3f8e74af3cbde96a1a471aa7