manifest(): Skip packages when using default system version
The loop which iterates over gcc, binutils and ${clibrary} expects the
sources of all these packages to be available so that we can copy
information into the manifest. However, this is not true when building
a single package (eg with --build gcc), leading to spurious error
messages in the logs, such as:

fatal: cannot change to 'default': No such file or directory
fatal: cannot change to 'log': No such file or directory

which happen because 'binutils' is defined to 'default' and 'clibrary'
to 'auto' in global.sh, leading to incorrect definition of srcdir.

This patch explicitly skips 'default' and '' values for srcdir.

Change-Id: I7caaf52f0b727e10fc15a322389062999b9b755f