lib/make.sh: Fix test-container tests on armhf
The glibc tests are built in two different modes (enabled at configure
time):

  1. No RPATH defined, tests will run by issuing the loader along with
     the required library paths. It is the default option.

  2. RPATH defined, tests will have RPATH pointing to the build
     directory, so the kernel will be responsible for running the
     loader.  It enables with --enable-hardcoded-path-in-tests.

In both cases, the loader consults the system cache (ld.so.cache), so
any compiler dynamic library (libgcc_s.so for instance) is supplied by
the system.

However, this might fail for container tests, where even test-container
(the binary that setups and runs the tests in a container mode) creates
a loader cache that might not contain the required libraries. The glibc
build system tries to check and provide the required libraries by
tracing the DT_NEEDED one from a test binary, but the tracing inhibits
the system loader, and thus it will not copy the required system
libraries. It is expected the user to copy the required libraries to
build a directory.

This works for case 2., however, test-container does not issue the
loader for case 1.: it will parse the command arguments and issue the
binary and the kernel will be responsible for starting the loader. This
has the side-effect of ignoring all the defined --library-path and thus
making the compiler libraries copy trick ineffective.

So we copy the compiler libraries on both the build directory and on the
sysroot used by the test-container to set each test. The former handles
the case 1., while the latter will place the libraries on the default
system path, which will be included in the loader cache as well (and
handling case 2.).

Change-Id: I5b90c733a92a5980a8f33e5b6a5c67f768f91906