lib/make.sh (make_docs): Build more docs
Build PDF docs for binutils, gdb and gcc.

Build man pages for binutils, gcc and libstdc++.
For libstdc++, this requires 'doxygen' and 'dot' to be present (they
are detected at configure time, and this is a no-op if they are
absent).

Note that gdbserver documentation is part of the gdb component.

Change-Id: I94579cf085bb2ff78f45c34b93015100ceb8ae0a