Add --force_build option
This new option works like --force, but forces only the build steps,
not the update of the sources.

This is a subset of --force, and is compatible with --disable update,
so that we can use it from CI for instance if we want to restart a
failed build but disabling --parallel in the second run, while also
making use of --enable maintainer_mode.

Change-Id: Ief1a492a1f70ba398d0442ecc160eb5c3869230a