Do not build a C++ compiler for stage1 targeting glibc
The glibc only requires a C++ compiler for testing. However, its
configure checks by default if a C++ compiler is present on PATH if
CXX is not set.

To avoid glibc configure fails with a stage1 compiler without C++
support, explicitly set CXX to an invalid value.  It will make the
configure stop searching on PATH for a possible C++ compiler
(which might have a target mismatch).

Change-Id: Id28e26d7d1e0ee2d481a431554db7fe01af1c701