Add and use qemu-wrapper.sh
This patch introduces scripts/qemu-wrapper.sh to invoke QEMU, allowing
more (pre)processing than the current one-liner in qemu.exp.

The initial motivation is to avoid errors at execution time in the GCC
testsuite related to LSan (leak sanitizer).

Indeed, LSan is enabled by default by ASan, but it is incompatible
with QEMU, so we want to disable it using the dedicated environment
variable.

Change-Id: Id74633c3acfb812ef7a78c08ee20d693e5c47d6f